backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
pydantic-settings
# library untuk database
sqlalchemy
psycopg2-binary
alembic
# Library Keamanan
passlib[bcrypt]
python-jose[cryptography]